requirements.txt
pytest==6.2.2
requests==2.25.1
allure-pytest==2.8.40